QcomModulePkg: Change memory allocations on LE NAND targets
When flashing images in fastboot, we only allocate 75% of the largest
contiguous piece of memory. However, this is not sufficient for LE NAND
devices with lower DDR sizes, so this change increases it to 85%. This
avoids sparsing of ubifs images, which is not supported by the fastboot
host. This requires that we decrease the usb buffer size though so we
don't use too much heap space.

Change-Id: I43d41d2ece45fe7d1167d20fba2958e9d194a24b