edk2: QcomModulePkg: Define Base versions for DDR APIs.
If GetDDRInfo Revision < EFI_DDRGETINFO_PROTOCOL_REVISION,
then ABL doesnt proceed further to evaluate ddr regions &
devicetype. This is not a fixed comparision as
EFI_DDRGETINFO_PROTOCOL_REVISION keeps on incrementing for
further versions. To Fix this issue, we need to declare macros
assigned with the version of DDR APIs in which they are introduced.
DdrInfoIf Base Version is 5. SCTConfig Base version is 7

Change-Id: I1f9d0832318d023cf6c722edf408c1db44faafe0